@@ -54,8 +54,8 @@
     "react-image-crop": "^9.1.1",
     "react-number-format": "^4.8.0",
     "react-redux": "^7.2.0",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
     "react-webcam": "^6.0.1",
     "redux": "^4.0.5",
@@ -65,7 +65,7 @@
     "sass": "^1.26.5",
     "sass-loader": "^10.0.2",
     "susy": "^2.2.14",
-    "typescript": "^3.7.5",
+    "typescript": "^4.8.4",
     "yup": "^0.29.0"
   },
   "scripts": {
