@@ -1,21 +1,17 @@
 import React, { FunctionComponent } from "react";
-import { Redirect, Route, RouteProps, useLocation } from "react-router-dom";
+import { Navigate, Route, RouteProps, useLocation } from "react-router-dom";
 import { isAuthenticated } from "../../../libraries/authUtils/isAuthenticated";
 import { useAuthentication } from "../../../libraries/authUtils/useAuthentication";
 
-const PrivateRoute: FunctionComponent<RouteProps> = ({ children, ...rest }) => {
+const PrivateRoute: FunctionComponent<RouteProps> = (props) => {
   const location = useLocation();
   useAuthentication();
+  console.log(location.pathname);
 
   return isAuthenticated() ? (
-    <Route {...rest}>{children}</Route>
+    <Route {...props} />
   ) : (
-    <Redirect
-      to={{
-        pathname: "/login",
-        state: { from: location },
-      }}
-    />
+    <Navigate to={location} replace />
   );
 };
 
