@@ -2,7 +2,7 @@ import isEmpty from "lodash.isempty";
 import React, { FunctionComponent, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { connect, useSelector } from "react-redux";
-import { Redirect, useParams } from "react-router";
+import { Navigate, useParams } from "react-router";
 import checkIcon from "../../../assets/check-icon.png";
 import { PATHS } from "../../../consts";
 import { PatientDTO } from "../../../generated";
@@ -41,6 +41,12 @@ const EditPatientActivity: FunctionComponent<TProps> = ({
   const { id } = useParams<{ id: string }>();
   const { t } = useTranslation();
 
+  useEffect(() => {
+    if (isEmpty(patient.data) && patient.status === "IDLE") {
+      getPatientThunk(id!);
+    }
+  }, [patient, id, getPatientThunk]);
+
   const breadcrumbMap = {
     [t("nav.patients")]: PATHS.patients,
     [t("nav.searchpatient")]: PATHS.patients_search,
@@ -76,13 +82,13 @@ const EditPatientActivity: FunctionComponent<TProps> = ({
 
   useEffect(() => {
     if (isEmpty(patient.data) && patient.status === "IDLE") {
-      getPatientThunk(id);
+      getPatientThunk(id!);
     }
   }, [patient, id, getPatientThunk]);
 
   useEffect(() => {
     if (activityTransitionState === "TO_PATIENT") {
-      getPatientThunk(id);
+      getPatientThunk(id!);
       updatePatientReset();
       setShouldResetForm(true);
     } else if (activityTransitionState === "TO_KEEP_EDITING") {
@@ -112,7 +118,7 @@ const EditPatientActivity: FunctionComponent<TProps> = ({
   switch (activityTransitionState) {
     case "TO_PATIENT":
       return (
-        <Redirect to={`${PATHS.patients_details}/${patient.data?.code}`} />
+        <Navigate to={`${PATHS.patients_details}/${patient.data?.code}`} replace />
       );
     default:
       return (
