@@ -1,34 +1,31 @@
 import React, { FC } from "react";
-import { Switch, useRouteMatch } from "react-router-dom";
-import PrivateRoute from "../../components/accessories/privateRoute/PrivateRoute";
+import { Route, Routes } from "react-router-dom";
 import DashboardActivity from "../../components/activities/dashboardActivity/DashboardActivity";
+import EditPatientActivity from "../../components/activities/editPatientActivity/EditPatientActivity";
 import NewPatientActivity from "../../components/activities/newPatientActivity/NewPatientActivity";
 import NotFound from "../../components/activities/notFound/NotFound";
 import SearchPatientActivity from "../../components/activities/searchPatientActivity/SearchPatientActivity";
+import { PATHS } from "../../consts";
 import { PatientDetailsRoutes } from "./PatientsDetailsRoutes";
 
-export const PatientsRoutes: FC = () => {
-  const { path } = useRouteMatch();
-  return (
-    <Switch>
-      <PrivateRoute exact path={path}>
+export const PatientsRoutes: FC = () => (
+  <Routes>
+    <Route
+      index
+      element={
         <DashboardActivity
-          newPatientRoute={`${path}/new`}
-          searchPatientRoute={`${path}/search`}
+          newPatientRoute={PATHS.patients_new}
+          searchPatientRoute={PATHS.patients_search}
         />
-      </PrivateRoute>
-      <PrivateRoute path={`${path}/new`}>
-        <NewPatientActivity dashboardRoute={path} />
-      </PrivateRoute>
-      <PrivateRoute path={`${path}/search`}>
-        <SearchPatientActivity />
-      </PrivateRoute>
-      <PrivateRoute path={`${path}/details/:id`}>
-        <PatientDetailsRoutes />
-      </PrivateRoute>
-      <PrivateRoute>
-        <NotFound backRoute={path} />
-      </PrivateRoute>
-    </Switch>
-  );
-};
+      }
+    />
+    <Route
+      path="new"
+      element={<NewPatientActivity dashboardRoute={PATHS.patients} />}
+    />
+    <Route path="search" element={<SearchPatientActivity />} />
+    <Route path="details/:id/*" element={<PatientDetailsRoutes />} />
+    <Route path="details/:id/edit" element={<EditPatientActivity />} />
+    <Route path="*" element={<NotFound />} />
+  </Routes>
+);
