@@ -10,8 +10,8 @@ import InfoBox from "../../infoBox/InfoBox";
 import { renderDate } from "../../../../libraries/formatUtils/dataFormatting";
 interface IOwnProps {
   shouldUpdateTable: boolean;
-  handleEdit: <T>(row: T) => void;
-  handleAddOperation: <T>(row: T) => void;
+  handleEdit: (row: any) => void;
+  handleAddOperation: (row: any) => void;
 }
 
 const PatientOPDTable: FunctionComponent<IOwnProps> = ({
