@@ -3,7 +3,7 @@ import set from "lodash.set";
 import { AUTH_KEY } from "../../consts";
 import { Middleware, RequestArgs } from "../../generated";
 import { SessionStorage } from "../storage/storage";
-import history from "../../history";
+import { redirect } from "react-router-dom";
 import { tokenHasExpired } from "../authUtils/tokenHasExpired";
 
 export const applyTokenMiddleware: Middleware = {
@@ -11,7 +11,7 @@ export const applyTokenMiddleware: Middleware = {
     const userCredentials = SessionStorage.read(AUTH_KEY);
     if (userCredentials.token && tokenHasExpired(userCredentials.token)) {
       SessionStorage.clear();
-      history.push("/login");
+      return redirect("/login");
     }
     return produce(request, (draft) => {
       if (userCredentials.token) {
